feat(pr/merge): add standalone `tsk pr merge` subcommand
Extract merge logic from `tsk done` into a reusable
`merge_pr_workflow` in pr.rs so the done workflow is fully
reproducible via individual commands.

- Add PrSubcommand::Merge with merge-method, auto-merge, yes, timeout
- Move merge helpers (confirm, wait, blocker detection) to pr.rs
- Rewire `tsk done` to call the shared merge_pr_workflow
- Add CLI help test for `tsk pr merge`
- Update README with manual step-by-step equivalent

Refs: #20
