+++ src/commands/done.rs
@@ -1,11 +1,8 @@
-use crate::adapters::backend::{BackendPrRecord, BackendProvider, MergeMethod, PrChecksStatus};
 use crate::adapters::git::{CliGit, GitBackend};
-use crate::adapters::prompts::{DialoguerPrompts, PromptBackend};
 use crate::cli::DoneArgs;
 use crate::commands::branch::{current_repo, cwd_utf8, find_issue_for_branch};
 use crate::commands::pr;
 use crate::config::load_config;
-use crate::domain::issue::IssueDocument;
 use crate::error::RiptskError;
 use crate::paths::AppPaths;
 use crate::services::auto_commit::maybe_auto_commit;
@@ -13,7 +10,6 @@ use crate::services::backend_mapping::{build_provider_for_backend, resolve_git_a
 use crate::services::id_resolution;
 use crate::services::issue_service::IssueService;
 use crate::storage::{frontmatter, issue_store};
-use std::time::{Duration, Instant};
 
 pub async fn run(paths: &AppPaths, args: DoneArgs) -> Result<(), RiptskError> {
     paths.require_initialized()?;
@@ -56,44 +52,13 @@ pub async fn run(paths: &AppPaths, args: DoneArgs) -> Result<(), RiptskError> {
         ));
     }
 
-    let initial_pr = provider.get_pr(repo_name, pr_number).await?;
-    let method = args.merge_method.unwrap_or_default();
-
-    if initial_pr.merged || initial_pr.state == "merged" {
-        crate::ui::info("PR already merged, continuing with cleanup");
-    } else {
-        if !args.yes
-            && !confirm_done(
-                &DialoguerPrompts,
-                &issue,
-                &initial_pr,
-                method,
-                args.auto_merge,
-            )?
-        {
-            crate::ui::warn("aborted");
-            return Ok(());
-        }
-
-        if args.auto_merge {
-            provider
-                .enable_auto_merge(repo_name, pr_number, method)
-                .await?;
-            crate::ui::info("Auto-merge enabled, waiting for checks...");
-            let _ = wait_for_merge(provider.as_ref(), repo_name, pr_number, args.timeout).await?;
-        } else if let Err(error) = provider
-            .merge_pr(repo_name, pr_number, method, None, None)
-            .await
-        {
-            if looks_like_checks_blocker(&error.to_string()) {
-                return Err(RiptskError::General(format!(
-                    "{error}. Try `tsk done {} --auto-merge`.",
-                    issue.frontmatter.id
-                )));
-            }
-            return Err(error);
-        }
-    }
+    let merge_opts = pr::MergeOptions {
+        merge_method: args.merge_method.unwrap_or_default(),
+        auto_merge: args.auto_merge,
+        yes: args.yes,
+        timeout: args.timeout,
+    };
+    pr::merge_pr_workflow(provider.as_ref(), repo_name, pr_number, &issue, &merge_opts).await?;
 
     let default_branch = provider.default_branch(repo_name).await?;
     git.checkout(repo_dir.as_path(), &default_branch)?;
@@ -135,118 +100,7 @@ pub async fn run(paths: &AppPaths, args: DoneArgs) -> Result<(), RiptskError> {
     Ok(())
 }
 
-fn confirm_done(
-    prompts: &dyn PromptBackend,
-    issue: &IssueDocument,
-    pr: &BackendPrRecord,
-    method: MergeMethod,
-    auto_merge: bool,
-) -> Result<bool, RiptskError> {
-    let mode = if auto_merge {
-        "enable auto-merge for"
-    } else {
-        "merge"
-    };
-    let prompt = format!(
-        "{} '{}' via {}?\n{}",
-        mode,
-        issue.frontmatter.title,
-        merge_method_label(method),
-        pr.url
-    );
-    prompts.confirm(&prompt, false)
-}
-
-async fn wait_for_merge(
-    provider: &dyn BackendProvider,
-    repo: &str,
-    pr_number: u64,
-    timeout_secs: u64,
-) -> Result<BackendPrRecord, RiptskError> {
-    let started = Instant::now();
-    loop {
-        let pr = provider.get_pr(repo, pr_number).await?;
-        if pr.merged || pr.state == "merged" {
-            return Ok(pr);
-        }
-
-        let checks = provider.get_pr_checks_status(repo, pr_number).await?;
-        crate::ui::info(&format!(
-            "checks status: {}",
-            pr_checks_status_label(checks)
-        ));
-        if checks == PrChecksStatus::Failed {
-            return Err(RiptskError::General(format!(
-                "checks failed for PR #{pr_number}"
-            )));
-        }
-        if started.elapsed() >= Duration::from_secs(timeout_secs) {
-            return Err(RiptskError::General(format!(
-                "timed out after {timeout_secs} seconds waiting for PR #{pr_number} to merge"
-            )));
-        }
-        tokio::time::sleep(Duration::from_secs(10)).await;
-    }
-}
-
-fn merge_method_label(method: MergeMethod) -> &'static str {
-    match method {
-        MergeMethod::Merge => "merge",
-        MergeMethod::Squash => "squash",
-        MergeMethod::Rebase => "rebase",
-    }
-}
-
-fn pr_checks_status_label(status: PrChecksStatus) -> &'static str {
-    match status {
-        PrChecksStatus::None => "none",
-        PrChecksStatus::Pending => "pending",
-        PrChecksStatus::Passed => "passed",
-        PrChecksStatus::Failed => "failed",
-    }
-}
-
-fn looks_like_checks_blocker(message: &str) -> bool {
-    let lower = message.to_lowercase();
-    lower.contains("status check")
-        || lower.contains("required check")
-        || lower.contains("check pending")
-        || lower.contains("pipeline")
-}
-
 fn is_branch_not_found_error(error: &RiptskError) -> bool {
     let message = error.to_string().to_lowercase();
     message.contains("404") || message.contains("not found")
-}
-
-#[cfg(test)]
-mod tests {
-    use super::{looks_like_checks_blocker, merge_method_label, pr_checks_status_label};
-    use crate::adapters::backend::{MergeMethod, PrChecksStatus};
-
-    #[test]
-    fn labels_merge_method_values() {
-        assert_eq!(merge_method_label(MergeMethod::Merge), "merge");
-        assert_eq!(merge_method_label(MergeMethod::Squash), "squash");
-        assert_eq!(merge_method_label(MergeMethod::Rebase), "rebase");
-    }
-
-    #[test]
-    fn labels_check_status_values() {
-        assert_eq!(pr_checks_status_label(PrChecksStatus::None), "none");
-        assert_eq!(pr_checks_status_label(PrChecksStatus::Pending), "pending");
-        assert_eq!(pr_checks_status_label(PrChecksStatus::Passed), "passed");
-        assert_eq!(pr_checks_status_label(PrChecksStatus::Failed), "failed");
-    }
-
-    #[test]
-    fn detects_check_related_merge_errors() {
-        assert!(looks_like_checks_blocker(
-            "Required status check \"ci\" is expected."
-        ));
-        assert!(looks_like_checks_blocker(
-            "pipeline must succeed before merge"
-        ));
-        assert!(!looks_like_checks_blocker("merge conflict"));
-    }
 }
